+++ src/client/lib/auth.ts
